requirements.txt
openai==1.59.6
openpyxl==3.1.3
pandas==1.5.3
pydantic==2.10.5
pydantic_settings==2.7.1
python-dotenv==1.0.1
tenacity==8.2.3
